requirements.txt
bcrypt==4.0.1
blinker==1.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.1.0
numpy==1.23.4
pandas==1.5.1
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
urllib3==1.26.13
Werkzeug==2.3.4
